refactor(app): extract help-buffer orchestrators (phase 5f)
Move open_help_manual and open_keyboard_shortcuts out of
buffer_management.rs into a new app/help_actions.rs sibling. ~107 lines.

Both open read-only virtual buffers populated with help content
(manual or keyboard shortcuts) and reuse an existing buffer if one
already exists. They're help-flow orchestrators, not buffer
management proper.

buffer_management.rs goes from 745 -> 635 lines.
2026 tests still pass.
